backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
pydantic==1.10.12
python-dotenv==1.0.0
requests==2.31.0
